Add graph labels
