+++ frontend/src/components/MoleculeSelector.tsx
@@ -2,30 +2,14 @@ import React, { useEffect, useState } from "react";
 import { TextField } from "@material-ui/core";
 import { Autocomplete } from "@material-ui/lab";
 import { CalcSpectrumParams } from "../constants";
+import {
+  addSubscriptsToMolecule,
+  removeSubscriptsFromMolecule,
+} from "../utils";
 
 interface MoleculesResponseData {
   molecules: string[];
 }
-
-const SUBSCRIPTS = "₁₂₃₄₅₆₇₈₉".split("");
-
-const addSubscriptsToMolecule = (molecule: string): string => {
-  return molecule
-    .split("")
-    .map((char) => (/^\d+$/.test(char) ? SUBSCRIPTS[parseInt(char) - 1] : char))
-    .join("");
-};
-
-const removeSubscriptsFromMolecule = (molecule: string): string => {
-  return molecule
-    .split("")
-    .map((char) =>
-      SUBSCRIPTS.includes(char)
-        ? (SUBSCRIPTS.indexOf(char) + 1).toString()
-        : char
-    )
-    .join("");
-};
 
 interface MoleculeSelectorProps {
   params: CalcSpectrumParams;
